Feat: 선택된 item props 로 받기
