+++ src/components/ItemMoveButton/ItemMoveButton.js
@@ -12,11 +12,10 @@ import { leftListState, rightListState } from '../../atom/objectAtom';
 import { useRecoilState } from 'recoil';
 import { emojiMenus } from '../../assets/data';
 
-const ItemMoveButton = () => {
+const ItemMoveButton = ({ selectedItems, setSelectedItems }) => {
   const [leftList, setLeftList] = useRecoilState(leftListState);
   const [rightList, setRightList] = useRecoilState(rightListState);
 
-  const [selectedItems, setSelectedItems] = useState([]);
   const handleInitializationClick = () => {
     setLeftList(emojiMenus);
     setRightList([]);
@@ -56,65 +55,28 @@ const ItemMoveButton = () => {
       setRightList([]);
     }
   };
-  const handleSelected = (e, option) => {
-    if (e.shiftKey) {
-      const targetItems = e.target.parentElement.id;
-      if (targetItems === 'left') {
-        setSelectedItems(leftList);
-      } else if (targetItems === 'right') {
-        setSelectedItems(rightList);
-      }
-    } else {
-      if (selectedItems.includes(option)) {
-        setSelectedItems((selected) => selected.filter((item) => item.id !== option.id));
-      } else {
-        setSelectedItems((selected) => [...selected, option]);
-      }
-    }
-  };
 
   return (
-    <FakeContainer>
-      <ul id="left">
-        {leftList.map((option) => (
-          <Item key={option.id} selected={selectedItems.includes(option)} onClick={(e) => handleSelected(e, option)}>
-            {option.name}
-          </Item>
-        ))}
-      </ul>
-      <Wrap>
-        <Button>
-          <FontAwesomeIcon icon={faArrowRotateRight} onClick={handleInitializationClick} />
-        </Button>
-        <Button>
-          <FontAwesomeIcon icon={faAngleDoubleLeft} onClick={() => handleMoveAll('right')} />
-        </Button>
-        <Button>
-          <FontAwesomeIcon icon={faAngleDoubleRight} onClick={() => handleMoveAll('left')} />
-        </Button>
-        <Button>
-          <FontAwesomeIcon icon={faAngleLeft} onClick={() => handleMoveSelected(selectedItems)} />
-        </Button>
-        <Button>
-          <FontAwesomeIcon icon={faAngleRight} onClick={() => handleMoveAvailable(selectedItems)} />
-        </Button>
-      </Wrap>
-      <ul id="right">
-        {rightList.map((option) => (
-          <Item key={option.id} selected={selectedItems.includes(option)} onClick={(e) => handleSelected(e, option)}>
-            {option.name}
-          </Item>
-        ))}
-      </ul>
-    </FakeContainer>
+    <Wrap>
+      <Button>
+        <FontAwesomeIcon icon={faArrowRotateRight} onClick={handleInitializationClick} />
+      </Button>
+      <Button>
+        <FontAwesomeIcon icon={faAngleDoubleLeft} onClick={() => handleMoveAll('right')} />
+      </Button>
+      <Button>
+        <FontAwesomeIcon icon={faAngleDoubleRight} onClick={() => handleMoveAll('left')} />
+      </Button>
+      <Button>
+        <FontAwesomeIcon icon={faAngleLeft} onClick={() => handleMoveSelected(selectedItems)} />
+      </Button>
+      <Button>
+        <FontAwesomeIcon icon={faAngleRight} onClick={() => handleMoveAvailable(selectedItems)} />
+      </Button>
+    </Wrap>
   );
 };
 export default ItemMoveButton;
-
-const FakeContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-`;
 
 const Wrap = styled.div`
   display: flex;
@@ -129,7 +91,4 @@ const Button = styled.button`
   border-radius: 3px;
   width: 30px;
   height: 30px;
-`;
-const Item = styled.li`
-  background-color: ${({ selected }) => (selected ? 'lightGrey' : 'white')};
 `;
